==> 00_GC_Data_Exploration.Rmd <==
---
title: "GC Data Exploration"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  # to simultaneously be run as chunks within the console
  # keep this set to inline: a bug in rstudio does not allow python and r
  chunk_output_type: inline
---

# Libraries

```{r}
library(tidyverse)  # CRAN v1.3.2 
library(plotly)     # CRAN v4.10.1 
library(reticulate) # CRAN v1.28 c("#525252", "#808080", "#E6E6E6")
library(lubridate)  # CRAN v1.9.2 
```

# Process GC Data

Run this code if new gc data has been added

```{r}
#system("python3 source/process_gc_data.py")
```

# Load data

```{r}
# Load GC raw data
gc <- read_csv("data/GC_Data_Table.csv") |> 
  select(-c(`...1`)) |> # remove this trash column
  rename(gc_sample_id = sample_type) |> 
  # fix mislabeled gas name
  mutate(
    gas = case_when(
      gas == "C02" ~ "CO2",
      TRUE ~ gas
    ),
    # fix incorrectly entered date formats
    file_name = str_replace(file_name, pattern = "Sept", replacement = "SEP"),
    file_name = str_replace(file_name, pattern = "Apr", replacement = "APR"),
    # Extract the analysis date from file name
    analysis_date = str_extract(file_name, "\\d{2}[A-Z]{3}\\d{2}"),
    # Convert to date
    analysis_date = as_date(analysis_date, format = "%d%b%y")
  )
  

# load sample_matrix metadata
sample_matrix <- read_csv("data/sample_matrix.csv") |> 
  # find number of days elapsed at GC timepoint
  mutate(
    inc_start_date = as.Date(inc_start_date, format = "%d-%b-%y"),
    GC_T1 = as_date(GC_T1, format = "%d-%b-%y"),
    GC_T2 = as_date(GC_T2, format = "%d-%b-%y"),
    GC_T3 = as_date(GC_T3, format = "%d-%b-%y"),
    GC_T4 = as_date(GC_T4, format = "%d-%b-%y"),
    inc_end_date = as.Date(inc_end_date, format = "%d-%b-%y"),
    saline_Ti = as_date(saline_Ti, format = "%d-%b-%y"),
    saline_Tf = as_date(saline_Tf, format = "%d-%b-%y"),
    frozen = as_date(frozen, format = "%d-%b-%y"),
    decanted = as_date(decanted, format = "%d-%b-%y"),
  ) |>
  mutate(
    dt1 = difftime(GC_T1, inc_start_date, units = "days"),
    dt2 = difftime(GC_T2, inc_start_date, units = "days"),
    dt3 = difftime(GC_T3, inc_start_date, units = "days"),
    dt4 = difftime(GC_T4, inc_start_date, units = "days"),
  )


## why this code is broken, I have no idea:
# across(
#       # List the columns to mutate across:
#       c(
#         inc_start_date, GC_T1, GC_T2,
#         GC_T3, GC_T4, inc_end_date, 
#         saline_Ti, saline_Tf, frozen, decanted
#         ),
#       # Define the lambda function ~ where .x is the across variable:
#       ~ as.Date(format = "%d-%b-%y"), 
#     )

# load list of bad gc standards injections
bad_gc_list <- readxl::read_excel("data/bad_gc_list.xlsx") |> pull(file_name)
```

# Data Cleanup

## Separate samples/standards

```{r}
# Create new tibble for just standard runs
gc_standards <- gc |> 
  filter(str_detect(gc_sample_id, "mix")) |> 
  mutate(
    sample_volume = as.numeric(str_extract(gc_sample_id, "\\d+(\\.\\d+)?(?=ml)"))
  ) |> 
  filter(sample_volume != 1)
```

## Parse text from sample code

```{r}
  
# Create new tibble for just sample runs
gc_samples <- gc |> 
  # filter for samples where first letter is
  filter(
      str_sub(gc_sample_id, 1, 1) == "p",
      !file_name %in% bad_gc_list
    ) |> 
  mutate(
    # extract sample volume
    sample_volume = as.numeric(str_extract(gc_sample_id, "\\d+(\\.\\d+)?(?=ml)")),
    # extract time point
    timepoint = str_extract(gc_sample_id, "T[0-9]"),
    # extract the incubation duration
    inc_time.d = str_extract(gc_sample_id, "(?<=\\D)\\d{1,3}(?=d)"),
    # extract the temperature
    temp = case_when(
      str_detect(gc_sample_id, "12C") ~ 12,
      str_detect(gc_sample_id, "-4C") ~ -4,
      str_detect(gc_sample_id, "4C") ~ 4
    ),
    # string version of temp
    temp.str = str_c(as.character(temp), "C"),
    # extract the site id
    site = case_when(
      str_detect(gc_sample_id, "pUS") ~ "US - 10cm",
      str_detect(gc_sample_id, "p83") ~ "83m",
      str_detect(gc_sample_id, "p54") ~ "54m",
      str_detect(gc_sample_id, "p35") ~ "35m",
    )
  ) |> 
  # clean up samples that I forgot to add timepoint code `T#`
  mutate(
    timepoint = case_when(
      # All samples on 28 and 13 OCT were T1!
      str_detect(file_name, "28OCT") & is.na(timepoint) ~ "T1",
      str_detect(file_name, "13OCT") & is.na(timepoint) ~ "T1",
      TRUE ~ timepoint
    ),
    gc_sample_id = case_when(
      str_detect(file_name, "28OCT") & is.na(timepoint) ~ str_c(gc_sample_id, "_", timepoint),
      str_detect(file_name, "13OCT") & is.na(timepoint) ~ str_c(gc_sample_id, "_", timepoint),
      TRUE ~ gc_sample_id
    )
  ) |> 
  # match sample_id for left_join
  mutate(
    sample_id = str_extract(gc_sample_id, "^[^_]*"),
    # fix the sample_ids that were not input correctly
    sample_id = case_when(
      gc_sample_id == "p83m(4C)_30d_T4_0.25ml" ~ "p83m(4C)30d",
      gc_sample_id == "p54m(4C)_30d_T4_0.5ml" ~ "p54m(4C)30d",
      gc_sample_id == "p83m(12C)_30d_T4_0.25ml" ~ "p83m(12C)30d",
      gc_sample_id == "p54m(12C)_30d_T4_0.25ml" ~ "p54m(12C)30d",
      gc_sample_id == "p35m(4C)_30d_T4_0.5ml" ~ "p35m(4C)30d",
      gc_sample_id == "p35m(12C)_30d_T4_0.25ml" ~ "p35m(12C)30d",
      gc_sample_id == "p35m(12C)_30d_T4_0.5ml" ~ "p35m(12C)30d",
      TRUE ~ sample_id
    )
  ) |> 
  # fix sample ids that don't have "m" after the site
  mutate(
    sample_id = case_when(
      str_detect(sample_id, "m", negate = TRUE) & str_detect(sample_id, "US", negate = TRUE) ~ 
        str_replace_all(sample_id, "(p\\d{2,})", "\\1m"),
      TRUE ~ sample_id
    )
  )
```

## Join metadata with gc data

```{r}
small_metadata <- sample_matrix |> 
  select(sample_id, dt1, dt2, dt3, dt4)

gc_samples_w_dt <- gc_samples |> 
  left_join(small_metadata, by = "sample_id") |> 
  # assign the appropriate dt for this gc measurement
  mutate(
    dt = case_when(
      timepoint == "T1" ~ dt1,
      timepoint == "T2" ~ dt2,
      timepoint == "T3" ~ dt3,
      timepoint == "T4" ~ dt4
    ),
    dt.d = parse_number(as.character(dt))
  ) |> 
  # remove extraneous dt columns
  select(-c(dt1, dt2, dt3, dt4))

gc_samples_w_dt_mean <- gc_samples_w_dt |> 
  # take mean value of gc injections
  group_by(site, temp, inc_time.d, timepoint, sample_id, gas) |> 
  summarize(
    .groups = "keep",
    # keep these columns:
    peak_area = mean(peak_area, na.rm = TRUE),
    peak_height = mean(peak_height, na.rm = TRUE),
    retention_time = mean(retention_time, na.rm = TRUE),
    correction_factor = mean(correction_factor, na.rm = TRUE),
    dt.d = mean(dt.d),
    sample_volume = sample_volume
  ) |> 
  mutate(
    temp.str = as.factor(str_c(temp, "C")), 
    temp.str = fct_relevel(temp.str, c("-4C", "4C", "12C"))
    )
  
gc_samples_w_dt <- gc_samples_w_dt |> 
  mutate(
    temp.str = as.factor(str_c(temp, "C")), 
    temp.str = fct_relevel(temp.str, c("-4C", "4C", "12C"))
    )
```

# Generate standard curves

Here we calculate the number of moles injected in our standard curve as follows:

$$
n = \frac{PV}{RT}
$$ where n is the number of moles, P is pressure in `atm`, V is volume in `mL`, R is the gas constant $R = 0.082057338 \space L\cdot atm \cdot K^{-1} mol^{-1}$.

```{r}
volume_to_moles <- function(pressure_atm = 0.8, vol_mL, R = 0.082057338	, temp_C = 20) {
  # n = PV/RT
  # R = 0.082057338	L.atm.K-1.mol-1
  # average atmospheric pressure in boulder is about 0.8 atm
  temp_K = temp_C + 273 # convert ˚C to ˚K
  vol_L = vol_mL / 1000 # convert mL to L
  n_moles = (pressure_atm * vol_L) / (R * temp_K)
  return(n_moles)
}

volume_to_moles(pressure_atm = 0.8, vol_mL = 0.5, temp_C = 20) # 1.6637e-05


gc_standard_curve <- gc_standards |>
  ungroup() |> 
  relocate(gas, peak_area, sample_volume) |> 
  filter(
    !is.na(sample_volume),
    !file_name %in% bad_gc_list
    ) |> 
  select(-c(retention_time, correction_factor, peak_height, gc_sample_id, column, carrier)) |> 
  mutate(
    mol_total = volume_to_moles(vol_mL = sample_volume), # moles of gas injected
    nmol_total = mol_total / 1E-9, # nmol gas injected (total),
    nmol_specific = nmol_total * .01, # 1% of the gas either CO2 or CH4
  )

# Tidy the fits
CH4_curve_fit <- lm(data = filter(gc_standard_curve, gas == "CH4"), peak_area ~ nmol_specific) |> broom::tidy() |> 
  mutate(gas = "CH4") |> 
  select(-c(std.error, statistic, p.value)) |> 
  pivot_wider(names_from = term, values_from = estimate)
CH4_curve_slope <- CH4_curve_fit |> pull(nmol_specific)
CH4_curve_int <- CH4_curve_fit |> pull(`(Intercept)`)
CH4_curve_equation <- paste("y =", round(CH4_curve_slope, 2), "+", round(CH4_curve_int, 2))

CO2_curve_fit <- lm(data = filter(gc_standard_curve, gas == "CO2"), peak_area ~ nmol_specific) |> 
  broom::tidy() |> 
  mutate(gas = "CO2") |> 
  select(-c(std.error, statistic, p.value)) |> 
  pivot_wider(names_from = term, values_from = estimate)
CO2_curve_slope <- CO2_curve_fit |> pull(nmol_specific)
CO2_curve_int <- CO2_curve_fit |> pull(`(Intercept)`)
CO2_curve_equation <- paste("y =", round(CO2_curve_slope, 2), "+", round(CO2_curve_int, 2))

CH4_curve_equation
CO2_curve_equation
```

Now we have a calibration that we can use to convert peak area to partial pressure:

$$
P = \frac{nRT}{V}
$$ but first we convert from peak area to $n$ nanomoles of gas injected:

$$
\begin{aligned}
A = nx + b \\
n = \frac{(A-b)}{x}
\end{aligned}
$$ where A is peak area, x is the defined slope, b is the y intercept, and n is the number of moles.

```{r}
peak_area_to_moles_per_mL <- function(peak_area, vol_mL, slope, intercept) {
  n_nmol = (peak_area - intercept) / slope # n = (A-b) / x
  mol = n_nmol / 1E9 # convert nanomoles to moles
  mol_per_mL = mol / vol_mL # convert to moles per mL gas
  return(
    case_when(
      mol_per_mL > 0 ~ mol_per_mL,
      mol_per_mL <= 0 ~ 0
    )
  )
}

peak_area_to_pressure <- function(peak_area, R = 0.082057338, temp_C = 20, vol_mL, slope, intercept) {
  # R = 0.082057338	L.atm.K-1.mol-1
  temp_K = 273 + temp_C # convert to ˚K
  vol_L = vol_mL / 1000 # convert to L
  n_nmol = (peak_area - intercept) / slope # n = (A-b) / x
  n_mol = n_nmol / 1E9 # convert nanomoles to moles
  pressure_atm = (n_mol * R * temp_K)/vol_L # P = nRT/V
  return(pressure_atm)
}

# Test it:
peak_area_to_pressure(25000, vol_mL = 0.5, slope = CO2_curve_slope, intercept = CO2_curve_int)
# 0.009287613 # should be about 1%
```

## Plot the standard curves

```{r}
p_std_curve <- gc_standard_curve |> 
  filter(gas != "CO") |> 
  ggplot() +
  aes(
    x = nmol_specific, 
    y = peak_area, 
    group = analysis_date,
    label = file_name
    ) +
  geom_point() +
  
  geom_abline(
    data = CH4_curve_fit,
    aes(
      slope = nmol_specific,
      intercept = `(Intercept)`
    ),
    color = "blue",
    linetype = "dashed"
    ) +
  geom_abline(
    data = CO2_curve_fit,
    aes(
      slope = nmol_specific,
      intercept = `(Intercept)`
    ),
    color = "blue",
    linetype = "dashed"
    ) +
  
  ggsci::scale_color_igv() +
  facet_wrap(vars(gas), scales = "free_y") +
  #coord_cartesian(ylim = c(10000, 50000)) +
  #scale_x_continuous(breaks = c(0.25, 0.5, 0.75, 1)) +
  theme_bw() +
  labs(
    color = "Gas",
    x = "nmol gas injected",
    y = "Peak Area"
  ) +
  theme(
    strip.background = element_blank(),
    legend.position = "bottom",
    panel.grid = element_blank()
  )

p_std_curve
```

# Convert peak area to moles/pressure

We calculate the number of moles per mL of gas, then the number of µg per mL of gas The MW of $CH_4$ is 16.04 g/mol. The MW of CO2 is 44.01 g/mol.

$$
\begin{aligned}
g/mL &= mol * MW \\
µg/mL &= g/mL * (1e^-6)
\end{aligned}
$$

```{r}
MW.CO2 <- 44.01 #g/mol
MW.CH4 <- 16.04 #g/mol


gc_samples_processed <- gc_samples_w_dt |> 
  filter(gas != "CO") |> 
  mutate(
    # add the standard curve parameters
    fit_slope = case_when(
      gas == "CO2" ~ CO2_curve_slope,
      gas == "CH4" ~ CH4_curve_slope
    ),
    fit_int = case_when(
      gas == "CO2" ~ CO2_curve_int,
      gas == "CH4" ~ CH4_curve_int
    ),
    # Calculate moles per mL of gas
    mol_per_ml = peak_area_to_moles_per_mL(
      peak_area = peak_area, 
      vol_mL = sample_volume, 
      slope = fit_slope, 
      intercept = fit_int
      ),
    nmol_per_ml = mol_per_ml * 1E9,
    partial_P_atm = peak_area_to_pressure(
      peak_area = peak_area,
      vol_mL = sample_volume,
      slope = fit_slope,
      intercept = fit_int
    ),
    # Calculate mass per mL of gas
    ug_per_ml = case_when(
      gas == "CO2" ~ (mol_per_ml * MW.CO2) * 1e-6,
      gas == "CH4" ~ (mol_per_ml * MW.CH4) * 1e-6
    )
  )
```

# Exploratory plots

## Peak area

```{r}
gc_samples_processed |> 
  filter(gas == "CH4") |> 
  #filter(inc_time.d != 180) |> 
  ggplot(
    aes(x = dt.d,
        y = peak_area / sample_volume,
        group = sample_id,
        color = as.factor(temp),
        shape = inc_time.d
        )
  ) +
  geom_point() +
  geom_line() +
  facet_wrap(vars(site), scale = "free_y") +
  scale_color_viridis_d(end = 0.9, option = "magma")
```

## Timeseries nmol

```{r}
p_co2 <- gc_samples_processed |> 
  filter(inc_time.d == 180) |> 
  filter(gas == "CO2") |> 
  ggplot() +
  aes(
    x = dt.d,
    y = nmol_per_ml,
    group = sample_id,
    label = file_name,
    color = as.factor(temp)
  ) +
  geom_point(
  ) +
  scale_shape_manual(values = c(0, 1)) +
  geom_line() +
  scale_colour_viridis_d(begin = 0.2, end = 0.8, option = "magma") +
  facet_wrap(vars(site), nrow = 1) +
  coord_cartesian(ylim = c(0, 900)) +
  theme_bw() +
  labs(
    x = latex2exp::TeX("$\\Delta$t (Days)"),
    y = latex2exp::TeX("$nmol \\cdot ml^{-1}$"),
    color = "Incubation Temp (C)"
  ) +
  theme(
    panel.grid = element_blank()
  )
p_co2

p_ch4 <- gc_samples_processed |> 
  filter(inc_time.d == 30) |> 
  filter(gas == "CH4") |> 
  ggplot() +
  aes(
    x = dt.d,
    y = nmol_per_ml,
    group = sample_id,
    label = file_name,
    color = as.factor(temp)
  ) +
  geom_point(
  ) +
  scale_shape_manual(values = c(0, 1)) +
  geom_line() +
  scale_colour_viridis_d(begin = 0.2, end = 0.8, option = "magma") +
  facet_wrap(vars(site), nrow = 1) +
  coord_cartesian(ylim = c(0, 25)) +
  theme_bw() +
  labs(
    x = latex2exp::TeX("$\\Delta$t (Days)"),
    y = latex2exp::TeX("$nmol \\cdot ml^{-1}$"),
    color = "Incubation Temp (C)"
  ) +
  theme(
    panel.grid = element_blank()
  )
p_ch4

p_co2ch4 <- cowplot::plot_grid(p_co2, p_ch4, nrow = 2)
cowplot::save_plot(
  filename = "fig_output/p_co2ch4.pdf",
  plot = p_co2ch4,
  base_height = 6,
  base_width = 10)
#ggplotly()
```
